feat: 스터디 생성 라우팅 경로 정리
